Move ft command to interface layer
